- Integrated DDA with wave generator - Changed the type of HP of creatures to double

+++ src/models/creatures/Aigle.java
@@ -46,7 +46,7 @@ public class Aigle extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public Aigle(long santeMax, int nbPiecesDOr, double vitesse)
+	public Aigle(double santeMax, int nbPiecesDOr, double vitesse)
 	{
 		this(0, 0, santeMax, nbPiecesDOr,vitesse);
 	}
@@ -64,7 +64,7 @@ public class Aigle extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public Aigle(int x, int y, long santeMax, int nbPiecesDOr, double vitesse)
+	public Aigle(int x, int y, double santeMax, int nbPiecesDOr, double vitesse)
 	{
 		super(x, y, 16, 16, santeMax, nbPiecesDOr, vitesse, 
 		      Creature.TYPE_AERIENNE, IMAGE, Constants.AIGLE);
