+++ src/models/creatures/MoutonNoir.java
@@ -48,7 +48,7 @@ public class MoutonNoir extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public MoutonNoir(long santeMax, int nbPiecesDOr, double vitesse)
+	public MoutonNoir(double santeMax, int nbPiecesDOr, double vitesse)
 	{
 		this(0, 0, santeMax, nbPiecesDOr,vitesse);
 	}
@@ -67,7 +67,7 @@ public class MoutonNoir extends Creature
 	 * @param nbPiecesDOr le nombre de pieces d'or de cette creature
 	 * @param vitesse vitesse de la creature
 	 */
-	public MoutonNoir(int x, int y, long santeMax, int nbPiecesDOr, double vitesse)
+	public MoutonNoir(int x, int y, double santeMax, int nbPiecesDOr, double vitesse)
 	{
 		super(x, y, 16, 16, santeMax, nbPiecesDOr, vitesse,
 		        Creature.TYPE_TERRIENNE, IMAGE, Constants.MOUTON_NOIR);
